Basic security Spring Boot MYSQL based authorization. Working example ready for implementation. Basic controller only. User management included, DTO added in user management Database created and relations established REST almost finished Error handling in progress

+++ src/main/java/com/sarpio/product/controller/CategoryController.java
@@ -1,26 +1,19 @@
 package com.sarpio.product.controller;
 
-import com.sarpio.exception.ApiRequestException;
 import com.sarpio.product.controller.dto.CategoryDto;
 import com.sarpio.product.service.CategoryService;
-import com.sun.net.httpserver.Headers;
-import io.swagger.models.Response;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.function.ServerRequest;
 
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
-@RequestMapping("/Category")
+@RequestMapping("/category")
 @RequiredArgsConstructor
 public class CategoryController {
 
@@ -28,28 +21,21 @@ public class CategoryController {
 
     @PostMapping("/")
     public ResponseEntity saveNewCategory(@Valid @RequestBody CategoryDto dto) {
-        CategoryDto categoryDto = null;
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        dto.setId(null);
-        try {
-            categoryDto = categoryService.saveNewCategory(dto);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return new ResponseEntity(headers, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-//        return categoryService.saveNewCategory(dto);
-        return new ResponseEntity(categoryDto, headers, HttpStatus.OK);
+        categoryService.saveNewCategory(dto);
+        return ResponseEntity.ok("New Category Saved!");
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteCategory(@PathVariable("id") @Min(1) Long id) {
+    public ResponseEntity deleteCategory(@Valid @PathVariable("id") @Min(1) Long id) {
+        if (categoryService.getCategoryById(id) == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         categoryService.deleteCategory(id);
-        return ResponseEntity.ok().build();
+        return new ResponseEntity<>("Deleted category with Id: " + id, HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public CategoryDto editCategory(@PathVariable("id") Long id, @Valid @RequestBody CategoryDto dto) {
+    public ResponseEntity editCategory(@PathVariable("id") Long id, @Valid @RequestBody CategoryDto dto) {
         return categoryService.editCategory(id, dto);
     }
