+++ src/main/java/com/sarpio/product/service/CategoryService.java
@@ -1,13 +1,14 @@
 package com.sarpio.product.service;
 
-import com.sarpio.exception.ApiRequestException;
 import com.sarpio.product.controller.dto.CategoryDto;
 import com.sarpio.product.model.CategoryEntity;
 import com.sarpio.product.repository.CategoryRepository;
 import com.sarpio.product.utils.EntityDtoMapper;
 import lombok.RequiredArgsConstructor;
-import net.bytebuddy.implementation.bytecode.Throw;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -17,24 +18,30 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
 
-    public CategoryDto saveNewCategory(CategoryDto dto) {
+    public ResponseEntity saveNewCategory(CategoryDto dto) {
         CategoryEntity entity = EntityDtoMapper.map(dto);
         categoryRepository.save(entity);
-        return EntityDtoMapper.map(entity);
+        return new ResponseEntity<CategoryDto>(EntityDtoMapper.map(entity), HttpStatus.OK);
     }
 
-    public void deleteCategory(Long id) {
-        CategoryEntity byId = categoryRepository.getById(id);
-        categoryRepository.deleteById(id);
+    public ResponseEntity deleteCategory(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } else {
+            categoryRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+
     }
 
-    public CategoryDto editCategory(Long id, CategoryDto dto) {
-        categoryRepository.findById(id)
-                .orElseThrow(() -> new ApiRequestException("There is no category with Id: " + id));
+    public ResponseEntity editCategory(Long id, CategoryDto dto) {
+        if (!categoryRepository.existsById(id)) {
+            return new ResponseEntity<>("Id: " + id + " not found", HttpStatus.NOT_FOUND);
+        }
         dto.setId(id);
         CategoryEntity save = EntityDtoMapper.map(dto);
         categoryRepository.save(save);
-        return EntityDtoMapper.map(save);
+        return new ResponseEntity(EntityDtoMapper.map(save), HttpStatus.OK);
     }
 
     public List<CategoryDto> listAllCategories() {
@@ -46,7 +53,7 @@ public class CategoryService {
 
     public CategoryDto getCategoryById(Long id) {
         CategoryEntity entity = categoryRepository.findById(id)
-                .orElseThrow(() -> new ApiRequestException("There is no category with Id: " + id));
+                .orElseThrow();
         return EntityDtoMapper.map(entity);
     }
 }
